Create QR.js
This code defines a function called `generateQRCode()`. When this function is called, it first retrieves the value of the input field with an `id` of "website". This is done using the `document.getElementById()` method and the value is stored in a variable called `website`.

The function then checks if the `website` variable has a truthy value (i.e. it's not an empty string, null, undefined, 0, false or NaN). If `website` has a truthy value, the function proceeds to generate a QR code using the `QRCode` library. The library takes two arguments: the element where the QR code should be displayed and the data to be encoded in the QR code. In this case, the element is a container with an `id` of "qrcode" and the data is the value of the `website` variable.

If `website` does not have a truthy value (i.e. it's an empty string, null, undefined, 0, false or NaN), the function displays an alert message asking the user to enter a valid URL.

Lastly, if the QR code is generated successfully, the function sets the `display` property of an element with an `id` of "qrcode-container" to "block", which makes the container visible on the page.